backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic>=2.0
pydantic-settings
email-validator
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pillow
anyio
httpx
mercadopago
stripe
pytest
